analysis_cn/requirements.txt
pandas>=1.5,<2.1
pyyaml>=6.0
jinja2>=3.1
matplotlib>=3.6,<3.8
seaborn>=0.13
scipy>=1.10,<1.11
numpy>=1.24,<1.25
click>=8.1
pyarrow>=10.0,<13.0
scikit-learn>=1.3,<1.4
openpyxl>=3.1
